Implement #better_insert
#insert will traverse the whole tree to get all values into an array,
add the new value, the rebuild the whole tree. The code is clean, but
maybe not efficient

#better_insert finds the node with the closest value then traverses
just its own children, adds the new value, and builds a subtree. Then
it finds the location of that node again and places a new subtree in
its place.

We could probably save an extra search by just storing the parent of
the closest node and directly working with node.left/right to
build a new tree and replace that same node.left/right